+++ modules/home-manager/services/default.nix
@@ -2,6 +2,7 @@
   imports = [
     ./gnome-keyring.nix
     ./picom.nix
+    ./polkit-agent.nix
     ./poweralertd.nix
     ./ssh-agent.nix
     ./xscreensaver.nix